scripts: Move download scripts to a directory.
They only ever get invoked from the main postinstall script.
