+++ scripts/download/lima.mjs
@@ -4,7 +4,7 @@ import childProcess from 'child_process';
 import fs from 'fs';
 import os from 'os';
 import path from 'path';
-import { download, getResource } from './download-resources.mjs';
+import { download, getResource } from '../download-resources.mjs';
 
 const limaRepo = 'https://github.com/rancher-sandbox/lima';
 const limaTag = 'v0.5.0';
